requirements.txt
openai>=1.12.0
anthropic>=0.18.0
PyGithub>=2.1.1
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
jsonschema>=4.20.0
gitpython>=3.1.40
